Add prefix sum executor
