+++ tests/python/test_sort.py
@@ -15,7 +15,7 @@ def test_sort():
                 values[i] = keys[i]
 
         fill()
-        ti._kernels.parallel_sort(keys, values)
+        ti.algorithms.parallel_sort(keys, values)
 
         keys_host = keys.to_numpy()
         values_host = values.to_numpy()
